jiuyu-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
python-dotenv==1.0.1
pydantic==2.6.4